requirements.txt
flask>=3.0
python-dotenv>=1.0
httpx>=0.27
feedparser>=6.0
pyyaml>=6.0
beautifulsoup4>=4.12
youtube-transcript-api>=0.6
anthropic>=0.40
openai>=1.50
